+++ R/compare.R
@@ -559,13 +559,16 @@ extract_slot <- function(x, i) if (.hasSlot(x, i)) slot(x, i) else missing_arg()
 path_slot <- function(path, i) glue::glue("{path}@{i}")
 compare_by_slot <- compare_by(index_slot, extract_slot, path_slot)
 
-index_prop <- function(x, y) union(S7::prop_names(x), S7::prop_names(y))
+# We compare S7 based not on their properties (which might have getters/setters)
+# but instead the underlying attributes that store any data. This might generate
+# confusing messages (e.g. if obj@x <- 1 actually sets property y)
+index_prop <- function(x, y) union(attr_names(x), attr_names(y))
 extract_prop <- function(x, i) {
   if (S7::prop_exists(x, i)) S7::prop(x, i) else missing_arg()
 }
 path_prop <- function(path, i) glue::glue("{path}@{i}")
 compare_by_prop <- compare_by(index_prop, extract_prop, path_prop)
-
+attr_names <- function(x) names(attributes(x))
 extract_fun <- function(x, i) switch(i, fn_body(x), fn_fmls(x), fn_env(x))
 path_fun <- function(path, i) {
   fun <- unname(c("body", "formals", "environment")[i])